Rename enums and move file to root dir
