+++ tibiapy/house.py
@@ -3,7 +3,7 @@ import re
 import bs4
 
 from tibiapy import Character, abc
-from tibiapy.enums.tibia_enums import Gender, HouseStatus
+from tibiapy.enums import Sex, HouseStatus
 from tibiapy.utils import parse_number_words, parse_tibia_datetime
 
 URL_HOUSE = "https://www.tibia.com/community/?subtopic=houses&page=view&houseid=%d&world=%s"
@@ -167,7 +167,7 @@ class House(abc.Serializable):
         if m:
             self.status = HouseStatus.RENTED
             self.owner = m.group("owner")
-            self.owner_sex = Gender.MALE if m.group("pronoun") == "He" else Gender.FEMALE
+            self.owner_sex = Sex.MALE if m.group("pronoun") == "He" else Sex.FEMALE
             self.paid_until = parse_tibia_datetime(m.group("paid_until"))
         else:
             self.status = HouseStatus.AUCTIONED
